move reportable association to a trait
