+++ app/Models/User.php
@@ -43,7 +43,7 @@ use Request;
 class User extends Model implements AuthenticatableContract
 {
     use Elasticsearch\UserTrait, Store\UserTrait;
-    use HasApiTokens, Authenticatable, UserAvatar, UserScoreable, Validatable;
+    use HasApiTokens, Authenticatable, Reportable, UserAvatar, UserScoreable, Validatable;
 
     protected $table = 'phpbb_users';
     protected $primaryKey = 'user_id';
@@ -1300,16 +1300,6 @@ class User extends Model implements AuthenticatableContract
         ]);
     }
 
-    public function reportBy(self $reporter, array $params = []) : UserReport
-    {
-        return $this->reportedIn()->create([
-            'comments' => $params['comments'] ?? '',
-            'reason' => $params['reason'] ?? 'Cheating',
-            'reporter_id' => $reporter->getKey(),
-            'user_id' => $this->getKey(),
-        ]);
-    }
-
     public function scopeDefault($query)
     {
         return $query->where([
@@ -1571,5 +1561,12 @@ class User extends Model implements AuthenticatableContract
         }
 
         return $this->isValid() && parent::save($options);
+    }
+
+    protected function newReportableExtraParams() : array
+    {
+        return [
+            'user_id' => $this->getKey(),
+        ];
     }
 }
